refactor(common/models): spin off model template based directly on LexiconTraversals
The new `TraversalModel` type inherits almost all code from `TrieModel`, serving as its base class.  That said, `TrieModel` will still be responsible for `Trie` loading.
